pretzel bank interest added

+++ kinokoV3/js/save.js
@@ -4,14 +4,19 @@ var fs = require("fs");
 exports.jsonSave = function(obj, name){
 	var json = JSON.stringify(obj);
 	var path = './json/' + name;
-
+	
+		
+	
 	if(fs.existsSync(path)){
 		var d = new Date();
 		var append = d.getTime();
-		var newPath = './json/old/' + append + '_' + name;
+		var oldFolder = './json/old' + d.getDate() +'/';
 		
-		fs.renameSync(path, newPath);
-		console.log(newPath + " has been created.");
+		if(fs.existsSync(oldFolder) || fs.mkdirSync(oldFolder)){
+			var newPath = oldFolder + append + '_' + name;
+			fs.renameSync(path, newPath);
+			console.log(newPath + " has been created.");
+		}
 	}
 	
 	fs.writeFile(path, json, 'utf8', (err) => {
